zephyr-flash-dfu.bbclass: add reset & retries support
Arduino-101 needs to be reset in order to enter "flash mode".
The flash window is about 5 seconds. If the flashing fails,
prompt the user to press reset.
After the device is flashed, it remains in the "flash mode"
until it is reset again, so it needs to enter the flash mode via reset
only once even if we flash images for multiple cores. So we only prompt
for reset once.

Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>